Apply PR comments
