+++ pinot-controller/src/main/java/org/apache/pinot/controller/api/resources/PauseStatus.java
@@ -1,0 +1,53 @@
+/**
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+package org.apache.pinot.controller.api.resources;
+
+import com.fasterxml.jackson.annotation.JsonCreator;
+import com.fasterxml.jackson.annotation.JsonInclude;
+import com.fasterxml.jackson.annotation.JsonProperty;
+import java.util.Set;
+
+
+@JsonInclude(JsonInclude.Include.NON_NULL)
+public class PauseStatus {
+  private boolean _pauseFlag;
+  private Set<String> _consumingSegments;
+  private String _description;
+
+  @JsonCreator
+  public PauseStatus(@JsonProperty("pauseFlag") boolean pauseFlag,
+      @JsonProperty("consumingSegments") Set<String> consumingSegments,
+      @JsonProperty("description") String description) {
+    _pauseFlag = pauseFlag;
+    _consumingSegments = consumingSegments;
+    _description = description;
+  }
+
+  public boolean getPauseFlag() {
+    return _pauseFlag;
+  }
+
+  public Set<String> getConsumingSegments() {
+    return _consumingSegments;
+  }
+
+  public String getDescription() {
+    return _description;
+  }
+}
